Fixing up the Grisu3 algorithm to better match the reference implementation (including comments, etc)

+++ src/System.Private.CoreLib/shared/System/Number.Dragon4.cs
@@ -36,7 +36,7 @@ namespace System
             // Extract meta data from the input double value.
             //
             // Refer to IEEE double precision floating point format.
-            ulong f = (ulong)(ExtractFractionAndBiasedExponent(value, out int e));
+            ulong f = ExtractFractionAndBiasedExponent(value, out int e);
             int mantissaHighBitIndex = (e == -1074) ? (int)(BigInteger.LogBase2(f)) : 52;
 
             // Step 2:
